[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.21 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/lv1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.cpp
@@ -5,5 +5,5 @@ using namespace std;
 
 vector<long long> solution(int x, int n) {
     vector<long long> answer;
-    for (int sum = x; n--; sum += x) { answer.push_back(sum); }
+    for (long long sum = x; n--; sum += x) { answer.push_back(sum); }
 return answer; }
